+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/AvatarConverterSettings.cs
@@ -19,7 +19,7 @@ namespace KRT.VRCQuestTools.Components
         /// </summary>
         [SerializeReference]
         public IMaterialConvertSettings defaultMaterialConvertSettings =
-            ToonStandardConvertSettings.SimpleFeatures;
+            ToonStandardConvertSettings.Default;
 
         /// <summary>
         /// Additional material convert settings.